audio_cutter/requirements.txt
faster-whisper
pydub
static-ffmpeg
google-genai
